Support MemConfs with very deep memories (#1367)

+++ src/main/scala/firrtl/passes/memlib/MemConf.scala
@@ -50,7 +50,7 @@ object MemConf {
 
   def fromString(s: String): Seq[MemConf] = {
     s.split("\n").toSeq.map(_ match {
-      case MemConf.regex(name, depth, width, ports, maskGran) => Some(MemConf(name, depth.toInt, width.toInt, MemPort.fromString(ports), Option(maskGran).map(_.toInt)))
+      case MemConf.regex(name, depth, width, ports, maskGran) => Some(MemConf(name, BigInt(depth), width.toInt, MemPort.fromString(ports), Option(maskGran).map(_.toInt)))
       case "" => None
       case _ => throw new Exception(s"Error parsing MemConf string : ${s}")
     }).flatten
